intercept download.osmand.net URIs
Intercept these URIs and translate them into direct links to OpenStreetMap
using HTTPS.  This is to ensure the least amount of data leakage possible.

Also included is the logic for creating OpenStreetMap short URLs, but on
Android, apps don't recognize that URI format, so its not used here at the
moment.