some location URLs cannot be parsed or fetched, so offer "Force HTTPS"
Some services use HTTP links to share location, but the location cannot be
parsed from the URL, nor cannot it be fetched from the site itself without
substantial code.  Glympse is one such service, since the shared URL is a
link to a person's location feed, which is Glympse's proprietary system.

closes #4401 https://dev.guardianproject.info/issues/4401