parse all Google Maps links, including short links
This first tries to parse the location from the URL itself, and if that
fails, it tries to first parse the location from the Redirect URL, then
it tries to get the JSON from Google, and parse that.  The JSON stuff
probably doesn't really work yet, but the rest does.

refs #4395 https://dev.guardianproject.info/issues/4395