create "Trusted App" chooser to selecting the map app to redirect to
To make this app more usable, this will give the user the option to set the
map app that Location Privacy should use to display incoming location URLs.
It also has the option for "Chooser" to show the standard Android chooser,
or "None" to just ignore location links entirely.

This is not hooked up to anything yet, it is just the UI.