when a URL is unparsable after all tricks, give Chooser or Blocked message
If the user has chosen to block all location links, then the chooser should
not be shown.  Otherwise, when parsing fails, offer the user a choice of
whether to try with an untrusted app.  Ultimately, there should be a pref
that lets the user decide the endgame when parsing a URL fails: either
block it, or put up a chooser.